Support checksum algorithm mismatch (+ other fixes) (#227)
* Consistent GPG ordering in lockfile

This prevents rpmoci updates that seem like the have changes, but are
just reordering the GPG entries.

* Remove debug print statement

* Support download if checksum algorithms don't match

The checksum algorithm (and hence checksum) can differ between resolving
the RPM and writing the lockfile, and downloading the RPM later. For
these cases, verify the checksum after downloading the RPM instead by
hashing the package ourselves with the recorded algorithm.

* Appease clippy

* release 0.4.4